Financial Connections: for v3, started modifying Partner Auth to support sheet styles.
